+++ src/dipdup/indexes/evm_subsquid_events/index.py
@@ -2,6 +2,7 @@ import asyncio
 import random
 import time
 from collections import defaultdict
+from collections import deque
 from typing import Any
 from typing import cast
 
@@ -21,9 +22,11 @@ from dipdup.models.evm_subsquid import SubsquidEvent
 from dipdup.models.evm_subsquid import SubsquidEventData
 from dipdup.models.evm_subsquid import SubsquidMessageType
 from dipdup.performance import metrics
+from dipdup.prometheus import Metrics
 
 LEVEL_BATCH_TIMEOUT = 1
 NODE_SYNC_LIMIT = 128
+NODE_BATCH_SIZE = 10
 
 
 class SubsquidEventsIndex(
@@ -99,7 +102,10 @@ class SubsquidEventsIndex(
                 break
 
         for message_level, level_logs in logs_by_level.items():
-            await self._process_level_events(tuple(level_logs), self.topics, message_level)
+            self._logger.info('Processing %s event logs of level %s', len(level_logs), message_level)
+            await self._process_level_events(tuple(level_logs), message_level)
+            if self._config.expose_metrics:
+                Metrics.set_sqd_processor_last_block(message_level)
 
     def get_sync_level(self) -> int:
         """Get level index needs to be synchronized to depending on its subscription status"""
@@ -131,6 +137,8 @@ class SubsquidEventsIndex(
             return
 
         subsquid_sync_level = await self.datasource.get_head_level()
+        if self._config.expose_metrics:
+            Metrics.set_sqd_processor_chain_height(subsquid_sync_level)
 
         use_node = False
         if self.node_datasources:
@@ -152,28 +160,52 @@ class SubsquidEventsIndex(
             for handler in self._config.handlers:
                 typename = handler.contract.module_name
                 topics.add(self.topics[typename][handler.name])
-            # FIXME: This is terribly inefficient (but okay for the last mile); see advanced example in web3.py docs.
-            for level in range(first_level, sync_level + 1):
-                # NOTE: Get random one every time
+
+            # NOTE: Requesting logs by batches of NODE_BATCH_SIZE.
+            batch_first_level = first_level
+            while batch_first_level <= sync_level:
+                batch_last_level = min(batch_first_level + NODE_BATCH_SIZE, sync_level)
                 level_logs = await self.random_node.get_logs(
                     {
-                        'fromBlock': hex(level),
-                        'toBlock': hex(level),
+                        'fromBlock': hex(batch_first_level),
+                        'toBlock': hex(batch_last_level),
                     }
                 )
-                block = await self.random_node.get_block_by_level(level)
-                if block is None:
-                    raise FrameworkException(f'Block {level} not found')
-                timestamp = int(block['timestamp'], 16)
-                parsed_level_logs = tuple(EvmNodeLogData.from_json(log, timestamp) for log in level_logs)
-                await self._process_level_events(parsed_level_logs, self.topics, sync_level)
 
+                # NOTE: We need block timestamps for each level, so fetch them separately and match with logs.
+                timestamps: dict[int, int] = {}
+                tasks: deque[asyncio.Task[None]] = deque()
+
+                async def _fetch_timestamp(level: int, timestamps: dict[int, int]) -> None:
+                    block = await self.random_node.get_block_by_level(level)
+                    timestamps[level] = int(block['timestamp'], 16)
+
+                for level in range(batch_first_level, batch_last_level + 1):
+                    tasks.append(asyncio.create_task(_fetch_timestamp(level, timestamps)))
+
+                await asyncio.gather(*tasks)
+
+                parsed_level_logs = tuple(
+                    EvmNodeLogData.from_json(
+                        log,
+                        timestamps[int(log['blockNumber'], 16)],
+                    )
+                    for log in level_logs
+                )
+
+                await self._process_level_events(parsed_level_logs, sync_level)
+                if self._config.expose_metrics:
+                    Metrics.set_sqd_processor_last_block(level)
+
+                batch_first_level = batch_last_level + 1
         else:
             sync_level = min(sync_level, subsquid_sync_level)
             fetcher = self._create_fetcher(first_level, sync_level)
 
             async for _level, events in fetcher.fetch_by_level():
-                await self._process_level_events(events, self.topics, sync_level)
+                await self._process_level_events(events, sync_level)
+                if self._config.expose_metrics:
+                    Metrics.set_sqd_processor_last_block(_level)
 
         await self._exit_sync_state(sync_level)
 
@@ -201,7 +233,6 @@ class SubsquidEventsIndex(
     async def _process_level_events(
         self,
         events: tuple[SubsquidEventData | EvmNodeLogData, ...],
-        topics: dict[str, dict[str, str]],
         sync_level: int,
     ) -> None:
         if not events: